Moving this class into it's own file
